backend/requirements.txt
django==4.1.2
mysqlclient
django-crispy-forms
pillow
django-cors-headers
djangorestframework
